# Overview In this work we have fixed the bugs related to simulation viewer. This section where the users can control different aspects of the simulation is now named as workshop. It is opened in a new tab. All workshop components rests inside a container component named SimulationWorkshop. SimulationWorkshop connects to the Socket.io socket that is created by the simulair-simulation-manager that is updated and run during the initialization of each instance. This process is controlled by simulation initialization bash scripts.
SimulationWorkshop component includes another component named SimulationImage which recieves the image stream that is sent by simulation core and renders is. The centeralized socket object is passed to this component after the socket connection is established.
NOTE: there is a potential bug that can probably happen if we try to change the socket connection once it is initialized.

We have also simplified the general design for making the development of new feature easier.